% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bainancovabayesianWrapper.R
\name{BainAncovaBayesian}
\alias{BainAncovaBayesian}
\title{Bain ANCOVA}
\usage{
BainAncovaBayesian(
  data = NULL,
  version = "1",
  bayesFactorMatrix = FALSE,
  bayesFactorPlot = FALSE,
  covariates = list(types = list(), value = list()),
  credibleInterval = 0.95,
  dependent = list(types = list(), value = ""),
  descriptives = FALSE,
  descriptivesPlot = FALSE,
  fixedFactors = list(types = list(), value = ""),
  fraction = 1,
  model = "",
  plotHeight = 320,
  plotWidth = 480,
  seed = 730518,
  standardized = FALSE
)
}
\description{
Bain ANCOVA
}
