% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bainanovabayesianWrapper.R
\name{BainAnovaBayesian}
\alias{BainAnovaBayesian}
\title{Bain ANOVA}
\usage{
BainAnovaBayesian(
  data = NULL,
  version = "1",
  bayesFactorMatrix = FALSE,
  bayesFactorPlot = FALSE,
  credibleInterval = 0.95,
  dependent = list(types = list(), value = ""),
  descriptives = FALSE,
  descriptivesPlot = FALSE,
  fixedFactors = list(types = list(), value = ""),
  fraction = 1,
  model = "",
  plotHeight = 320,
  plotWidth = 480,
  seed = 353263,
  standardized = FALSE
)
}
\description{
Bain ANOVA
}
