% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bainregressionlinearbayesianWrapper.R
\name{BainRegressionLinearBayesian}
\alias{BainRegressionLinearBayesian}
\title{Regression}
\usage{
BainRegressionLinearBayesian(
  data = NULL,
  version = "1",
  bayesFactorMatrix = FALSE,
  bayesFactorPlot = FALSE,
  covariates = list(types = list(), value = list()),
  credibleInterval = 0.95,
  dependent = list(types = list(), value = ""),
  descriptives = FALSE,
  fraction = 1,
  model = "",
  plotHeight = 320,
  plotWidth = 480,
  seed = 836204,
  standardized = FALSE
)
}
\description{
Regression
}
