% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bainsembayesianWrapper.R
\name{BainSemBayesian}
\alias{BainSemBayesian}
\title{Bain Structural Equation Modeling}
\usage{
BainSemBayesian(
  data = NULL,
  version = "1",
  bayesFactorMatrix = FALSE,
  bayesFactorPlot = FALSE,
  credibleInterval = 0.95,
  descriptives = FALSE,
  factorStandardisation = "std.lv",
  fixedFactors = list(types = "unknown", value = ""),
  fraction = 1,
  model = "",
  pathDiagram = FALSE,
  pathDiagramEstimates = FALSE,
  pathDiagramLegend = FALSE,
  plotHeight = 320,
  plotWidth = 480,
  seed = 802582,
  standardized = FALSE,
  syntax = list(columns = list(), model = "", modelOriginal = "")
)
}
\description{
Bain Structural Equation Modeling
}
