% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bainttestbayesianindependentsamplesWrapper.R
\name{BainTTestBayesianIndependentSamples}
\alias{BainTTestBayesianIndependentSamples}
\title{Bain Welch's T-Test}
\usage{
BainTTestBayesianIndependentSamples(
  data = NULL,
  version = "1",
  bayesFactorPlot = FALSE,
  bayesFactorType = "BF01",
  credibleInterval = 0.95,
  descriptives = FALSE,
  descriptivesPlot = FALSE,
  fraction = 1,
  groupingVariable = list(types = list(), value = ""),
  hypothesis = "equalNotEqual",
  plotHeight = 320,
  plotWidth = 480,
  seed = 421018,
  variables = list(types = list(), value = list())
)
}
\description{
Bain Welch's T-Test
}
