% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bainttestbayesianonesampleWrapper.R
\name{BainTTestBayesianOneSample}
\alias{BainTTestBayesianOneSample}
\title{Bain One Sample T-Test}
\usage{
BainTTestBayesianOneSample(
  data = NULL,
  version = "1",
  bayesFactorPlot = FALSE,
  bayesFactorType = "BF01",
  credibleInterval = 0.95,
  descriptives = FALSE,
  descriptivesPlot = FALSE,
  fraction = 1,
  hypothesis = "equalNotEqual",
  plotHeight = 320,
  plotWidth = 480,
  seed = 972076,
  testValue = 0,
  variables = list(types = list(), value = list())
)
}
\description{
Bain One Sample T-Test
}
