% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bainttestbayesianpairedsamplesWrapper.R
\name{BainTTestBayesianPairedSamples}
\alias{BainTTestBayesianPairedSamples}
\title{Bain Paired Samples T-Test}
\usage{
BainTTestBayesianPairedSamples(
  data = NULL,
  version = "1",
  bayesFactorPlot = FALSE,
  bayesFactorType = "BF01",
  credibleInterval = 0.95,
  descriptives = FALSE,
  descriptivesPlot = FALSE,
  fraction = 1,
  hypothesis = "equalNotEqual",
  pairs = list(),
  plotHeight = 320,
  plotWidth = 480,
  seed = 188626
)
}
\description{
Bain Paired Samples T-Test
}
